+++ cedar-policy/src/api.rs
@@ -872,9 +872,10 @@ impl PartialResponse {
         self.0.concretize().into()
     }
 
-    /// Returns the set of [`PolicyId`]s that were definitely satisfied
-    pub fn definitely_satisfied(&self) -> impl Iterator<Item = &PolicyId> {
-        self.0.definitely_satisfied().map(PolicyId::ref_cast)
+    /// Returns the set of [`Policy`]s that were definitely satisfied.
+    /// This will be the set of policies (both `permit` and `forbid`) that evaluated to `true`
+    pub fn definitely_satisfied(&self) -> impl Iterator<Item = Policy> + '_ {
+        self.0.definitely_satisfied().map(Policy::from_ast)
     }
 
     /// Returns the set of [`PolicyId`]s that encountered errors
@@ -884,26 +885,29 @@ impl PartialResponse {
 
     /// Returns an over-approximation of the set of determining policies
     ///
-    /// This is all policies that may be determining for any substitution of the unknowns
-    pub fn may_be_determining(&self) -> impl Iterator<Item = &PolicyId> {
-        self.0.may_be_determining().map(PolicyId::ref_cast)
+    /// This is all policies that may be determining for any substitution of the unknowns.
+    /// Policies not in this set will not affect the final decision, regardless of any
+    /// substitutions.
+    ///
+    /// For more information on what counts as "determining" see: <https://docs.cedarpolicy.com/auth/authorization.html#request-authorization>
+    pub fn may_be_determining(&self) -> impl Iterator<Item = Policy> + '_ {
+        self.0.may_be_determining().map(Policy::from_ast)
     }
 
     /// Returns an under-approximation of the set of determining policies
     ///
-    /// This is all policies that must be determining for all possible substitutions of the unknowns
-    pub fn must_be_determining(&self) -> impl Iterator<Item = &PolicyId> {
-        self.0.must_be_determining().map(PolicyId::ref_cast)
+    /// This is all policies that must be determining for all possible substitutions of the unknowns.
+    /// This set will include policies that evaluated to `true` and are guaranteed to be
+    /// contributing to the final authorization decision.
+    ///
+    /// For more information on what counts as "determining" see: <https://docs.cedarpolicy.com/auth/authorization.html#request-authorization>
+    pub fn must_be_determining(&self) -> impl Iterator<Item = Policy> + '_ {
+        self.0.must_be_determining().map(Policy::from_ast)
     }
 
     /// Returns the set of non-trivial (meaning more than just `true` or `false`) residuals expressions
     pub fn nontrivial_residuals(&'_ self) -> impl Iterator<Item = Policy> + '_ {
         self.0.nontrivial_residuals().map(Policy::from_ast)
-    }
-
-    /// Returns the set of [`PolicyId`]s of non-trivial (meaning more than just `true` or `false`) residuals expressions
-    pub fn nontrivial_residual_ids(&self) -> impl Iterator<Item = &PolicyId> {
-        self.0.nontrivial_residual_ids().map(RefCast::ref_cast)
     }
 
     /// Returns every policy as a residual expression
@@ -912,8 +916,8 @@ impl PartialResponse {
     }
 
     /// Return the residual for a given [`PolicyId`], if it exists in the response
-    pub fn get(&self, id: &PolicyId) -> Option<est::Expr> {
-        self.0.get(&id.0).map(|e| est::Expr::from(e.clone()))
+    pub fn get(&self, id: &PolicyId) -> Option<Policy> {
+        self.0.get(&id.0).map(Policy::from_ast)
     }
 
     /// Attempt to re-authorize this response given a mapping from unknowns to values
@@ -1809,6 +1813,15 @@ impl SourceLocation {
     /// have a range.
     pub fn range_end(&self) -> Option<usize> {
         self.0.source_loc().map(parser::Loc::end)
+    }
+
+    /// Returns a tuple of (start, end) of the location.
+    /// Returns `None` if this location does not have a range.
+    pub fn range_start_and_end(&self) -> Option<(usize, usize)> {
+        self.0
+            .source_loc()
+            .as_ref()
+            .map(|loc| (loc.start(), loc.end()))
     }
 }
 
@@ -2749,7 +2762,7 @@ impl Template {
         self.ast.slots().map(|slot| SlotId::ref_cast(&slot.id))
     }
 
-    /// Get the head constraint on this policy's principal
+    /// Get the scope constraint on this policy's principal
     pub fn principal_constraint(&self) -> TemplatePrincipalConstraint {
         match self.ast.principal_constraint().as_inner() {
             ast::PrincipalOrResourceConstraint::Any => TemplatePrincipalConstraint::Any,
@@ -2780,7 +2793,7 @@ impl Template {
         }
     }
 
-    /// Get the head constraint on this policy's action
+    /// Get the scope constraint on this policy's action
     pub fn action_constraint(&self) -> ActionConstraint {
         // Clone the data from Core to be consistent with the other constraints
         match self.ast.action_constraint() {
@@ -2794,7 +2807,7 @@ impl Template {
         }
     }
 
-    /// Get the head constraint on this policy's resource
+    /// Get the scope constraint on this policy's resource
     pub fn resource_constraint(&self) -> TemplateResourceConstraint {
         match self.ast.resource_constraint().as_inner() {
             ast::PrincipalOrResourceConstraint::Any => TemplateResourceConstraint::Any,
@@ -2870,7 +2883,7 @@ impl FromStr for Template {
     }
 }
 
-/// Head constraint on policy principals.
+/// Scope constraint on policy principals.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum PrincipalConstraint {
     /// Un-constrained
@@ -2885,7 +2898,7 @@ pub enum PrincipalConstraint {
     IsIn(EntityTypeName, EntityUid),
 }
 
-/// Head constraint on policy principals for templates.
+/// Scope constraint on policy principals for templates.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum TemplatePrincipalConstraint {
     /// Un-constrained
@@ -2913,7 +2926,7 @@ impl TemplatePrincipalConstraint {
     }
 }
 
-/// Head constraint on policy actions.
+/// Scope constraint on policy actions.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum ActionConstraint {
     /// Un-constrained
@@ -2924,7 +2937,7 @@ pub enum ActionConstraint {
     Eq(EntityUid),
 }
 
-/// Head constraint on policy resources.
+/// Scope constraint on policy resources.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum ResourceConstraint {
     /// Un-constrained
@@ -2939,7 +2952,7 @@ pub enum ResourceConstraint {
     IsIn(EntityTypeName, EntityUid),
 }
 
-/// Head constraint on policy resources for templates.
+/// Scope constraint on policy resources for templates.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum TemplateResourceConstraint {
     /// Un-constrained
@@ -3100,7 +3113,7 @@ impl Policy {
         self.ast.is_static()
     }
 
-    /// Get the head constraint on this policy's principal
+    /// Get the scope constraint on this policy's principal
     pub fn principal_constraint(&self) -> PrincipalConstraint {
         let slot_id = ast::SlotId::principal();
         match self.ast.template().principal_constraint().as_inner() {
@@ -3123,7 +3136,7 @@ impl Policy {
         }
     }
 
-    /// Get the head constraint on this policy's action
+    /// Get the scope constraint on this policy's action
     pub fn action_constraint(&self) -> ActionConstraint {
         // Clone the data from Core to be consistant with the other constraints
         // INVARIANT: all of the EntityUids come from a policy, which must have Concrete EntityUids
@@ -3139,7 +3152,7 @@ impl Policy {
         }
     }
 
-    /// Get the head constraint on this policy's resource
+    /// Get the scope constraint on this policy's resource
     pub fn resource_constraint(&self) -> ResourceConstraint {
         let slot_id = ast::SlotId::resource();
         match self.ast.template().resource_constraint().as_inner() {
@@ -4571,7 +4584,7 @@ mod test {
     #[test]
     fn test_partition_fold() {
         let even_or_odd = |s: &str| {
-            i64::from_str_radix(s, 10).map(|i| {
+            i64::from_str(s).map(|i| {
                 if i % 2 == 0 {
                     Either::Left(i)
                 } else {
@@ -4605,12 +4618,12 @@ mod test {
     #[test]
     fn test_est_policyset_encoding() {
         let mut pset = PolicySet::default();
-        let policy: Policy = r#"permit(principal, action, resource) when { principal.foo };"#
+        let policy: Policy = r"permit(principal, action, resource) when { principal.foo };"
             .parse()
             .unwrap();
         pset.add(policy.new_id(PolicyId::new("policy"))).unwrap();
         let template: Template =
-            r#"permit(principal == ?principal, action, resource) when { principal.bar };"#
+            r"permit(principal == ?principal, action, resource) when { principal.bar };"
                 .parse()
                 .unwrap();
         pset.add_template(template.new_id(PolicyId::new("template")))